+++ android/src/main/java/me/carda/awesome_notifications/notifications/enumerators/NotificationLayout.java
@@ -8,6 +8,7 @@ public enum NotificationLayout {
     ProgressBar,
     Messaging,
     MessagingGroup,
-    MediaPlayer
+    MediaPlayer,
+    PhoneCall
 }