+++ example/lib/pages/home_page.dart
@@ -1,6 +1,7 @@
 import 'dart:io';
 import 'dart:math';
 
+import 'package:awesome_notifications/android_foreground_service.dart';
 import 'package:awesome_notifications_example/common_widgets/led_light.dart';
 import 'package:awesome_notifications_example/common_widgets/seconds_slider.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
@@ -35,6 +36,7 @@ class _HomePageState extends State<HomePage> {
 
   bool delayLEDTests = false;
   double _secondsToWakeUp = 5;
+  double _secondsToCallCategory = 5;
 
   bool globalNotificationsAllowed = false;
   bool schedulesFullControl = false;
@@ -167,6 +169,31 @@ class _HomePageState extends State<HomePage> {
     });
 
     AwesomeNotifications().actionStream.listen((receivedAction) {
+
+      if(receivedAction.channelKey == 'call_channel'){
+        switch (receivedAction.buttonKeyPressed) {
+
+          case 'REJECT':
+            AndroidForegroundService.stopForeground();
+            break;
+
+          case 'ACCEPT':
+            loadSingletonPage(targetPage: PAGE_PHONE_CALL, receivedAction: receivedAction);
+            AndroidForegroundService.stopForeground();
+            break;
+
+          default:
+            loadSingletonPage(targetPage: PAGE_PHONE_CALL, receivedAction: receivedAction);
+            break;
+        }
+        return;
+      }
+
+      if (receivedAction.channelKey == 'alarm_channel') {
+        AndroidForegroundService.stopForeground();
+        return;
+      }
+
       if (!StringUtils.isNullOrEmpty(receivedAction.buttonKeyInput)) {
         processInputTextReceived(receivedAction);
       } else if (!StringUtils.isNullOrEmpty(
@@ -184,6 +211,13 @@ class _HomePageState extends State<HomePage> {
           refreshPermissionsIcons();
       })
     );
+  }
+
+  void loadSingletonPage({required String targetPage, required ReceivedAction receivedAction}){
+    // Avoid to open the notification details page over another details page already opened
+    Navigator.pushNamedAndRemoveUntil(context, targetPage,
+            (route) => (route.settings.name != targetPage) || route.isFirst,
+        arguments: receivedAction);
   }
 
   Future<void> refreshPermissionsIcons() async {
@@ -227,23 +261,20 @@ class _HomePageState extends State<HomePage> {
     );
   }
 
-  void processDefaultActionReceived(ReceivedAction receivedNotification) {
+  void processDefaultActionReceived(ReceivedAction receivedAction) {
     Fluttertoast.showToast(msg: 'Action received');
 
     String targetPage;
 
     // Avoid to reopen the media page if is already opened
-    if (receivedNotification.channelKey == 'media_player') {
+    if (receivedAction.channelKey == 'media_player') {
       targetPage = PAGE_MEDIA_DETAILS;
       if (Navigator.of(context).isCurrent(PAGE_MEDIA_DETAILS)) return;
     } else {
       targetPage = PAGE_NOTIFICATION_DETAILS;
     }
 
-    // Avoid to open the notification details page over another details page already opened
-    Navigator.pushNamedAndRemoveUntil(context, targetPage,
-        (route) => (route.settings.name != targetPage) || route.isFirst,
-        arguments: receivedNotification);
+    loadSingletonPage(targetPage: targetPage, receivedAction: receivedAction);
   }
 
   void processInputTextReceived(ReceivedAction receivedAction) {
@@ -539,6 +570,25 @@ class _HomePageState extends State<HomePage> {
                 backgroundColor: Colors.red,
                 labelColor: Colors.white,
                 onPressed: () => NotificationUtils.cancelNotification(1)),
+
+            /* ******************************************************************** */
+
+            TextDivisor(title: 'Notification\'s Special Category'),
+            TextNote('The notification category is a group of predefined categories that best describe the nature of the notification and may be used by some systems for ranking, delay or filter the notifications. Its highly recommended to correctly categorize your notifications..\n\n'
+                'Slide the bar above to add some delay on notification.'),
+            SecondsSlider(steps: 12, minValue: 0, onChanged: (newValue){ setState(() => _secondsToCallCategory = newValue ); }),
+            SimpleButton('Show call notification',
+                onPressed: () => Future.delayed(Duration(seconds: _secondsToCallCategory.toInt()), () {
+                  NotificationUtils.showCallNotification(1);
+                })),
+            SimpleButton('Show alarm notification',
+                onPressed: () => Future.delayed(Duration(seconds: _secondsToCallCategory.toInt()), () {
+                  NotificationUtils.showAlarmNotification(1);
+                })),
+            SimpleButton('Stop Alarm / Call',
+                backgroundColor: Colors.red,
+                labelColor: Colors.white,
+                onPressed: () => NotificationUtils.stopForegroundServiceNotification()),
 
             /* ******************************************************************** */