+++ example/lib/utils/common_functions.dart
@@ -21,6 +21,14 @@ Future<String> saveAssetOnDisk(ImageProvider image, String fileName) async {
   return filePath;
 }
 
+String printDuration(Duration? duration) {
+  if (duration == null) return '00:00';
+  String twoDigits(int n) => n.toString().padLeft(2, "0");
+  String twoDigitMinutes = twoDigits(duration.inMinutes.remainder(60));
+  String twoDigitSeconds = twoDigits(duration.inSeconds.remainder(60));
+  return "$twoDigitMinutes:$twoDigitSeconds";
+}
+
 Future<String> downloadAndSaveImageOnDisk(String url, String fileName) async {
   var directory = await getApplicationDocumentsDirectory();
   var filePath = '${directory.path}/$fileName';